adding stocks and Controllers to stocks and users

+++ user/src/main/java/com/pedroblome/user/repository/UserRepository.java
@@ -1,16 +1,11 @@
 package com.pedroblome.user.repository;
 
-
 import com.pedroblome.user.model.User;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-
-
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
-    
+public interface UserRepository extends JpaRepository<User, Long> {
 
 }
-
